+++ src/assets/js/main.js
@@ -1,4 +1,5 @@
 import { initSearchMenu } from './search-menu';
+import { initLazyImages } from './lazy-images';
 
 async function loadFilterHandler() {
 	const input = document.querySelector( 'input#search_table_input' );
@@ -17,6 +18,7 @@ async function loadFilterHandler() {
  */
 const main = () => {
 	initSearchMenu();
+	initLazyImages();
 	loadFilterHandler();
 };